requirements.txt
openai>=1.0
anthropic>=0.30
google-generativeai>=0.5
pandas>=2.0
pydantic>=2.0
tenacity>=8.0
tqdm>=4.0
Pillow>=10.0
